requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.6
transformers==4.35.2
torch==2.1.1
pinecone-client==2.2.4
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.24.3
tiktoken==0.5.1
sentence-transformers==2.2.2 